QE-1039: Refactor API authentication
