--- application/libraries/Api/Auth/ApiAuthException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace LimeSurvey\Api\Auth;
-
-use LimeSurvey\Api\ApiException;
-
-class ApiAuthException extends ApiException
-{
-}
