+++ application/libraries/Api/Authentication/AuthenticationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace LimeSurvey\Api\Authentication;
+
+use LimeSurvey\Api\ApiException;
+
+class ApiAuthenticationException extends ApiException
+{
+}
